Remove unnecessary files and fix build
