+++ src/components/index.ts
@@ -1,4 +1,3 @@
 export * from './buttons';
 export * from './layout';
 export * from './indicators';
-export * from './other';
