+++ src/hooks/index.ts
@@ -1,1 +1,2 @@
 export * from './useMediaPicker';
+export * from './useMediaProcessor';
