--- src/utils/processors/index.ts
@@ -1,1 +1,0 @@
-export * from './useMediaProcessor';
